Refactor circular references
